Relocate ledger related execptions.
