+++ wrappers/java/src/test/java/org/hyperledger/indy/sdk/ledger/SchemaRequestsTest.java
@@ -2,7 +2,6 @@ package org.hyperledger.indy.sdk.ledger;
 
 import org.hyperledger.indy.sdk.IndyIntegrationTestWithPoolAndSingleWallet;
 import org.hyperledger.indy.sdk.InvalidStructureException;
-import org.hyperledger.indy.sdk.pool.InvalidLedgerTransactionException;
 import org.hyperledger.indy.sdk.signus.Signus;
 import org.hyperledger.indy.sdk.signus.SignusResults;
 import org.json.JSONObject;
